+++ checkpoint/orbax/checkpoint/path/step_test.py
@@ -15,6 +15,7 @@
 """Tests for Orbax step storage constructs."""
 
 from typing import Iterator
+from unittest import mock
 from absl.testing import absltest
 from absl.testing import parameterized
 from etils import epath
@@ -178,6 +179,7 @@ class StandardNameFormatTest(parameterized.TestCase):
       step,
       expected,
   ):
+    """Tests for committed steps with `find_step()` and `find_step_path()`."""
     for step_dir in self.generate_step_names():
       (self.directory / step_dir).mkdir(parents=True, exist_ok=True)
 
@@ -188,12 +190,97 @@ class StandardNameFormatTest(parameterized.TestCase):
     if expected is None:
       with self.assertRaises(ValueError):
         _ = name_format.find_step(self.directory, step)
-      return
+      with self.subTest('find_step_path_with_include_uncommitted=False'):
+        with self.assertRaises(ValueError):
+          _ = step_lib.find_step_path(
+              self.directory,
+              name_format,
+              step=step,
+              include_uncommitted=False,
+          )
+    else:
+      self.assertEqual(
+          self.directory / expected,
+          name_format.find_step(self.directory, step).path,
+      )
+      with self.subTest('uncommitted_gcs_step'):
+        with mock.patch.object(
+            step_lib, 'is_gcs_path', autospec=True, return_value=True
+        ):
+          with self.assertRaises(ValueError):
+            _ = name_format.find_step(self.directory, step)
+      with self.subTest('find_step_path_with_include_uncommitted=False'):
+        self.assertEqual(
+            self.directory / expected,
+            step_lib.find_step_path(
+                self.directory,
+                name_format,
+                step=step,
+                include_uncommitted=False,
+            ),
+        )
+      with self.subTest(
+          'find_uncommited_gcs_step_path_with_include_uncommitted=False'
+      ):
+        with mock.patch.object(
+            step_lib, 'is_gcs_path', autospec=True, return_value=True
+        ):
+          with self.assertRaises(ValueError):
+            _ = step_lib.find_step_path(
+                self.directory,
+                name_format,
+                step=step,
+                include_uncommitted=False,
+            )
 
-    self.assertEqual(
-        self.directory / expected,
-        name_format.find_step(self.directory, step).path,
-    )
+  @parameterized.parameters(True, False)
+  def test_find_step_path_with_uncommitted_checkpoint(self, gcs: bool):
+    """Tests for `step.find_step_path(include_uncommitted=True)`."""
+    name_format = step_lib.standard_name_format(step_prefix='step')
+    if gcs:
+      uncommitted_step_name = 'step_200'
+      for step_name in ['step_1', 'step_100', uncommitted_step_name]:
+        step_dir = self.directory / step_name
+        step_dir.mkdir(parents=True, exist_ok=True)
+        if step_name is not uncommitted_step_name:
+          commit_success_file = step_dir / step_lib._COMMIT_SUCCESS_FILE
+          commit_success_file.write_text(
+              f'Checkpoint commit was successful to {step_dir}'
+          )
+    else:
+      uncommitted_step_name = step_lib.get_tmp_directory(
+          step_lib.get_save_directory(
+              step=200, directory=self.directory, step_name_format=name_format
+          )
+      ).name
+      for step_name in ['step_1', 'step_100', uncommitted_step_name]:
+        (self.directory / step_name).mkdir(parents=True, exist_ok=True)
+
+    if gcs:
+      with mock.patch.object(
+          step_lib, 'is_gcs_path', autospec=True, return_value=True
+      ):
+        self.assertEqual(
+            self.directory / uncommitted_step_name,
+            step_lib.find_step_path(
+                self.directory, name_format, step=200, include_uncommitted=True
+            ),
+        )
+        with self.assertRaises(ValueError):
+          _ = step_lib.find_step_path(
+              self.directory, name_format, step=300, include_uncommitted=True
+          )
+    else:
+      self.assertEqual(
+          self.directory / uncommitted_step_name,
+          step_lib.find_step_path(
+              self.directory, name_format, step=200, include_uncommitted=True
+          ),
+      )
+      with self.assertRaises(ValueError):
+        _ = step_lib.find_step_path(
+            self.directory, name_format, step=300, include_uncommitted=True
+        )
 
   @parameterized.named_parameters(
       dict(
@@ -259,6 +346,14 @@ class StandardNameFormatTest(parameterized.TestCase):
             for metadata in name_format.find_all(self.directory)
         ]),
     )
+    with self.subTest('find_all_uncommitted_gcs_steps'):
+      with mock.patch.object(
+          step_lib, 'is_gcs_path', autospec=True, return_value=True
+      ):
+        self.assertEmpty([
+            metadata.path.name
+            for metadata in name_format.find_all(self.directory)
+        ])
 
 
 class UtilsTest(parameterized.TestCase):